--- src/main/java/sungminyun/springBoard/board/Board.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.board;
-
-public class Board {
-}
