--- src/main/java/sungminyun/springBoard/board/repository/BoardRepository.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.board.repository;
-
-public interface BoardRepository {
-}
