--- src/main/java/sungminyun/springBoard/board/service/BoardService.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.board.service;
-
-public interface BoardService {
-}
