--- src/main/java/sungminyun/springBoard/comment/Comment.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.comment;
-
-public class Comment {
-}
