--- src/main/java/sungminyun/springBoard/comment/repository/CommentRepository.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.comment.repository;
-
-public class CommentRepository {
-}
