--- src/main/java/sungminyun/springBoard/comment/service/CommentService.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.comment.service;
-
-public interface CommentService {
-}
