+++ src/main/java/sungminyun/springBoard/model/Board.java
@@ -1,0 +1,4 @@
+package sungminyun.springBoard.model;
+
+public class Board {
+}
