+++ src/main/java/sungminyun/springBoard/model/Comment.java
@@ -1,0 +1,4 @@
+package sungminyun.springBoard.model;
+
+public class Comment {
+}
