+++ src/main/java/sungminyun/springBoard/model/Post.java
@@ -1,0 +1,4 @@
+package sungminyun.springBoard.model;
+
+public class Post {
+}
