+++ src/main/java/sungminyun/springBoard/model/User.java
@@ -1,0 +1,4 @@
+package sungminyun.springBoard.model;
+
+public class User {
+}
