--- src/main/java/sungminyun/springBoard/post/Post.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.post;
-
-public class Post {
-}
