--- src/main/java/sungminyun/springBoard/post/repository/PostRepository.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.post.repository;
-
-public interface PostRepository {
-}
