--- src/main/java/sungminyun/springBoard/post/service/PostService.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.post.service;
-
-public interface PostService {
-}
