+++ src/main/java/sungminyun/springBoard/repository/BoardRepository.java
@@ -1,0 +1,4 @@
+package sungminyun.springBoard.repository;
+
+public interface BoardRepository {
+}
