+++ src/main/java/sungminyun/springBoard/service/CommentService.java
@@ -1,0 +1,4 @@
+package sungminyun.springBoard.service;
+
+public interface CommentService {
+}
