+++ src/main/java/sungminyun/springBoard/service/UserService.java
@@ -1,0 +1,4 @@
+package sungminyun.springBoard.service;
+
+public interface UserService {
+}
