--- src/main/java/sungminyun/springBoard/user/User.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.user;
-
-public class User {
-}
