--- src/main/java/sungminyun/springBoard/user/repository/UserRepository.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.user.repository;
-
-public interface UserRepository {
-}
