--- src/main/java/sungminyun/springBoard/user/service/UserService.java
@@ -1,4 +1,0 @@
-package sungminyun.springBoard.user.service;
-
-public interface UserService {
-}
